[Improve][Connector-V2] Refactor the package of local file connector
